Move StructMember centric functions to a new more appropriate StructMemberHelper class
Add boolean property isAttribute to StructMember

Implement generateResponseAttributesParser, Similar to generateResponseParser but handles attribute properties of an XML node rather than name / value of a child node.
