+++ ds3-autogen-c/src/main/resources/templates/source-templates/ResponseParser.ftl
@@ -1,16 +1,30 @@
 <#-- ******************************************************** -->
 <#-- Generate "TypedefStructResponseParser" from Structs      -->
-<#--   Input: Source object                                   -->
+<#--   Input: Struct object                                   -->
 <#-- ******************************************************** -->
 static ds3_error* _parse_${structEntry.getName()}(const ds3_client* client, const xmlDocPtr doc, const xmlNodePtr root, ${structEntry.getName()}** _response) {
+<#if structHelper.hasAttributes(structEntry)>
+    struct _xmlAttr* attribute;
+</#if>
+<#if structHelper.hasChildNodes(structEntry)>
     xmlNodePtr child_node;
+</#if>
     ${structEntry.getName()}* response;
     ds3_error* error = NULL;
 
     response = g_new0(${structEntry.getName()}, 1);
-    for (child_node = root->xmlChildrenNode; child_node != NULL; child_node = child_node->next) {
-${structHelper.generateResponseParser(structEntry.getStructMembers(), false)}
+
+<#if structHelper.hasAttributes(structEntry)>
+    for (attribute = root->properties; attribute != NULL; attribute = attribute->next) {
+${structMemberHelper.generateResponseAttributesParser(structEntry.getStructMembers())}
     }
+</#if>
+
+<#if structHelper.hasChildNodes(structEntry)>
+    for (child_node = root->xmlChildrenNode; child_node != NULL; child_node = child_node->next) {
+${structMemberHelper.generateResponseParser(structEntry.getStructMembers())}
+    }
+</#if>
 
     <#if structHelper.hasComplexMembers(structEntry)>
     if (error == NULL) {
