+++ ds3-autogen-c/src/main/resources/templates/source-templates/ResponseParserTopLevel.ftl
@@ -5,7 +5,12 @@
 static ds3_error* _parse_top_level_${structEntry.getName()}(const ds3_client* client, const ds3_request* request, ${structEntry.getName()}** _response, GByteArray* xml_blob) {
     xmlDocPtr doc;
     xmlNodePtr root;
+<#if structHelper.hasChildNodes(structEntry)>
     xmlNodePtr child_node;
+</#if>
+<#if structHelper.hasAttributes(structEntry)>
+    struct _xmlAttr* attribute;
+</#if>
     ${structEntry.getName()}* response;
     ds3_error* error = NULL;
 
@@ -15,9 +20,17 @@ static ds3_error* _parse_top_level_${structEntry.getName()}(const ds3_client* cl
     }
 
     response = g_new0(${structEntry.getName()}, 1);
-    for (child_node = root->xmlChildrenNode; child_node != NULL; child_node = child_node->next) {
-${structHelper.generateResponseParser(structEntry.getStructMembers(), false)}
+<#if structHelper.hasAttributes(structEntry)>
+    for (attribute = root->properties; attribute != NULL; attribute = attribute->next) {
+${structMemberHelper.generateResponseAttributesParser(structEntry.getStructMembers())}
     }
+</#if>
+
+<#if structHelper.hasChildNodes(structEntry)>
+    for (child_node = root->xmlChildrenNode; child_node != NULL; child_node = child_node->next) {
+${structMemberHelper.generateResponseParser(structEntry.getStructMembers())}
+    }
+</#if>
 
     xmlFreeDoc(doc);
