+++ ds3-autogen-c/src/test/java/com/spectralogic/ds3autogen/c/StructHelper_Test.java
@@ -30,6 +30,7 @@ import com.spectralogic.ds3autogen.c.converters.StructConverter;
 import com.spectralogic.ds3autogen.c.helpers.EnumHelper;
 import com.spectralogic.ds3autogen.c.helpers.RequestHelper;
 import com.spectralogic.ds3autogen.c.helpers.StructHelper;
+import com.spectralogic.ds3autogen.c.helpers.StructMemberHelper;
 import com.spectralogic.ds3autogen.c.models.Enum;
 import com.spectralogic.ds3autogen.c.models.*;
 import com.spectralogic.ds3autogen.utils.TestFileUtilsImpl;
@@ -51,8 +52,8 @@ public class StructHelper_Test {
     public void testTypeRequiresNewParser() {
         final Set<String> existingStructs = new HashSet<>();
         final ImmutableSet.Builder<String> enumNames = ImmutableSet.builder();
-        final StructMember testStruct1 = new StructMember(new PrimitiveType("int", false), "intMember");
-        final StructMember testStruct2 = new StructMember(new PrimitiveType("ds3_bool", false), "boolMember");
+        final StructMember testStruct1 = new StructMember(new PrimitiveType("int", false), "intMember", false);
+        final StructMember testStruct2 = new StructMember(new PrimitiveType("ds3_bool", false), "boolMember", false);
         final ImmutableList<StructMember> testStructMembers = ImmutableList.of(testStruct1, testStruct2);
         final Struct testStruct = new Struct("testStruct", "Data", testStructMembers, false, false, false);
         assertFalse(StructHelper.requiresNewCustomParser(testStruct, existingStructs, enumNames.build()));
@@ -62,8 +63,8 @@ public class StructHelper_Test {
     public void testTypeRequiresNewParser2() {
         final Set<String> existingStructs = new HashSet<>();
         final ImmutableSet.Builder<String> enumNames = ImmutableSet.builder();
-        final StructMember testStruct1 = new StructMember(new PrimitiveType("ds3_bool", false), "boolMember");
-        final StructMember testStruct2 = new StructMember(new FreeableType("ds3_user_api_bean_response", false), "beanMember");
+        final StructMember testStruct1 = new StructMember(new PrimitiveType("ds3_bool", false), "boolMember", false);
+        final StructMember testStruct2 = new StructMember(new FreeableType("ds3_user_api_bean_response", false), "beanMember", false);
         final ImmutableList<StructMember> testStructMembers = ImmutableList.of(testStruct1, testStruct2);
         final Struct testStruct = new Struct("testStruct", "Data", testStructMembers, false, false, false);
         assertTrue(StructHelper.requiresNewCustomParser(testStruct, existingStructs, enumNames.build()));
@@ -74,8 +75,8 @@ public class StructHelper_Test {
         final ImmutableSet.Builder<String> enumNames = ImmutableSet.builder();
         existingStructs.add("ds3_tape_type");
 
-        final StructMember testStruct1 = new StructMember(new PrimitiveType("ds3_bool", false), "boolMember");
-        final StructMember testStruct2 = new StructMember(new FreeableType("ds3_tape_type", false), "tapeTypeMember");
+        final StructMember testStruct1 = new StructMember(new PrimitiveType("ds3_bool", false), "boolMember", false);
+        final StructMember testStruct2 = new StructMember(new FreeableType("ds3_tape_type", false), "tapeTypeMember", false);
         final ImmutableList<StructMember> testStructMembers = ImmutableList.of(testStruct1, testStruct2);
         final Struct testStruct = new Struct("testStruct", "Data", testStructMembers, false, false, false);
         assertFalse(StructHelper.requiresNewCustomParser(testStruct, existingStructs, enumNames.build()));
@@ -90,7 +91,7 @@ public class StructHelper_Test {
 
         final ImmutableSet.Builder<String> enumNames = ImmutableSet.builder();
         final Struct testStruct = StructConverter.toStruct(ds3Type, enumNames.build(), ImmutableSet.of(), ImmutableSet.of());
-        final String output = StructHelper.generateStructMembers(testStruct.getStructMembers());
+        final String output = StructMemberHelper.generateStructMembers(testStruct.getStructMembers());
         assertTrue(output.contains("ds3_bool bool_element;"));
         assertTrue(output.contains("ds3_user_api_bean_response* bean_element;"));
     }
@@ -104,7 +105,7 @@ public class StructHelper_Test {
 
         final ImmutableSet.Builder<String> enumNames = ImmutableSet.builder();
         final Struct testStruct = StructConverter.toStruct(ds3Type, enumNames.build(), ImmutableSet.of(), ImmutableSet.of());
-        final String output = StructHelper.generateResponseParser(testStruct.getStructMembers(), false);
+        final String output = StructMemberHelper.generateResponseParser(testStruct.getStructMembers());
 
         assertTrue(output.contains("    if (element_equal(child_node, \"BoolElement\")) {"));
         assertTrue(output.contains("        response->bool_element = xml_get_bool(client->log, doc, child_node);"));
@@ -244,5 +245,25 @@ public class StructHelper_Test {
 
         assertTrue(output.contains("    return error;"));
         assertTrue(output.contains("}"));
+    }
+
+    @Test
+    public void testHasAttributes() {
+        final Struct testAttributesStruct = new Struct("testAttributesStruct", null,
+                ImmutableList.of(new StructMember(new PrimitiveType("int", false), "num_objects", true)),
+                false, false, false);
+        assertTrue(StructHelper.hasAttributes(testAttributesStruct));
+    }
+
+    @Test
+    public void testHasAttributesMixed() {
+        final Struct testAttributesStruct = new Struct("testAttributesStruct", null,
+                ImmutableList.of(
+                        new StructMember(new PrimitiveType("uint64_t", false), "some_attribute", true),
+                        new StructMember(new PrimitiveType("int", false), "num_objects", false),
+                        new StructMember(new FreeableType("ds3_object", true), "objects_node", false)),
+                false, false, false);
+        assertTrue(StructHelper.hasAttributes(testAttributesStruct));
+        assertTrue(StructHelper.hasChildNodes(testAttributesStruct));
     }
 }
